backend/requirements.txt
# Framework Web e Servidor
fastapi>=0.109.0
uvicorn>=0.27.0

# Banco de Dados (PostgreSQL e ORM)
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Segurança e Autenticação (JWT e Senhas)
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9

# Processamento de Dados (Upload da Planilha Excel)
pandas>=2.2.0
openpyxl>=3.1.2

# Inteligência Artificial (Google Gemini)
google-generativeai>=0.3.2

# Variáveis de Ambiente (Para ler o arquivo .env localmente)
python-dotenv>=1.0.1